src/aalibrary/requirements.txt
--index-url https://pypi.python.org/simple
echopype
echoregions; python_version>="3.10"
boto3
gcsfs
hvplot
panel
google-api-python-client
google-cloud
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
google-cloud-vision
numpy>=1.19.5
openpyxl>=3.1.3
pandas
pandas_gbq
pytest
pytz>=2024.1
tqdm
azure-storage-file-datalake>=12.7.0
zarr==2.8.3
inquirerpy
loguru
numcodecs==0.15.1; python_version>="3.11"
mkdocs
mkdocs-material
mkdocstrings-python
mkdocs-git-revision-date-localized-plugin
mkdocs-open-in-new-tab
nbformat
python-dotenv
sqlparse
google-genai>=0.3.0
tomli-w>=1.0.0
tomli>=2.0.0; python_version < "3.11"